+++ src/utils/register-ts-node.ts
@@ -1,0 +1,6 @@
+type TSNode = typeof import('ts-node')
+
+export function registerTSNode() {
+  const tsNode: TSNode = require('ts-node')
+  tsNode.register({})
+}